加强分包支持，Text表格, NumericType支持分包
NumericType改成const int，方便支持分包

+++ Packages/cn.etetet.wow/Scripts/Hotfix/Server/Map/NumericChange_DamageRemoveBuff.cs
@@ -3,7 +3,7 @@
 namespace ET.Server
 {
     [Module(ModuleName.Spell)]
-    [NumericWatcher(SceneType.Map, (int)NumericType.HP)]
+    [NumericWatcher(SceneType.Map, NumericType.HP)]
     public class NumericChange_DamageRemoveBuff: INumericWatcher
     {
         public void Run(Unit unit, NumbericChange args)
