+++ Packages/cn.etetet.wow/Scripts/HotfixView/Client/Unit/NumericWatcher_Speed_ChangeMotionSpeed.cs
@@ -3,7 +3,7 @@
 	/// <summary>
 	/// 客户端监视speed数值变化，改变血条值
 	/// </summary>
-	[NumericWatcher(SceneType.Current, (int)NumericType.Speed)]
+	[NumericWatcher(SceneType.Current, NumericType.Speed)]
 	public class NumericWatcher_Speed_ChangeMotionSpeed : INumericWatcher
 	{
 		public void Run(Unit unit, NumbericChange args)
